Refactor wgpu renderer
