+++ pygfx/renderers/svg/_svgrenderer.py
@@ -2,10 +2,11 @@ import os
 import io
 import pylinalg as la
 
-from .._base import Renderer, RenderFunctionRegistry
+from ...utils.renderfunctionregistry import RenderFunctionRegistry
 
 from ...objects import WorldObject
 from ...cameras import Camera
+from .. import Renderer
 
 
 registry = RenderFunctionRegistry()
