+++ pygfx/renderers/wgpu/shaders/__init__.py
@@ -1,0 +1,10 @@
+""" All pygfx builtin shaders.
+"""
+
+from . import backgroundshader
+from . import meshshader
+from . import pointsshader
+from . import lineshader
+from . import imageshader
+from . import volumeshader
+from . import textshader
