+++ pygfx/renderers/wgpu/shaders/imageshader.py
@@ -1,10 +1,18 @@
 import wgpu  # only for flags/enums
 
-from . import register_wgpu_render_function, WorldObjectShader, Binding, RenderMask
-from ._utils import to_texture_format, GfxSampler, GfxTextureView
-from ...objects import Image
-from ...materials import ImageBasicMaterial
-from ...resources import Texture
+from ....objects import Image
+from ....materials import ImageBasicMaterial
+from ....resources import Texture
+
+from .. import (
+    register_wgpu_render_function,
+    WorldObjectShader,
+    Binding,
+    RenderMask,
+    to_texture_format,
+    GfxSampler,
+    GfxTextureView,
+)
 
 
 vertex_and_fragment = wgpu.ShaderStage.VERTEX | wgpu.ShaderStage.FRAGMENT
