+++ pygfx/utils/renderfunctionregistry.py
